backend/requirements.txt
# === Core ===
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic[email]>=2.0
pydantic-settings>=2.0

# === Database ===
sqlalchemy>=2.0
asyncpg>=0.30
alembic>=1.14
redis>=5.0

# === ML ===
scikit-learn>=1.5
xgboost>=2.1
numpy>=1.26
pandas>=2.2
joblib>=1.3
torch>=2.4
shap>=0.45

# === HTTP / Async ===
httpx>=0.27

# === Auth / Security ===
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.12
webauthn>=2.2
pyotp>=2.9
qrcode[pil]>=8.0

# === Testing ===
pytest>=8.0
pytest-asyncio>=0.24
httpx>=0.27

# === Monitoring ===
sentry-sdk[fastapi]>=2.0

# === Utilities ===
python-dotenv>=1.0
aioredis>=2.0
